Create settings service and inject into account settings
