+++ src/app/pages/accountsettings/account-setting.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from "@angular/core";
+import { SettingsService } from "src/app/services/settings.service";
 
 @Component({
     selector: 'app-account-settings',
@@ -6,41 +7,18 @@ import { Component, OnInit } from "@angular/core";
 })
 
 export class AccountSettingsComponent implements OnInit{
-
-    //Get element from index with theme as ID
-    public linkTheme = document.querySelector('#theme');
-    //Get elements with selector class
-    public links!: NodeListOf<Element>;
+    //Move all my code to the service file and inject on constructor
+    constructor(private settingsService: SettingsService){}
 
     ngOnInit(): void{
-        this.links = document.querySelectorAll('.selector');
-        this.checkCurrentTheme();
+        //Indica en que theme esta actualmente
+        this.settingsService.checkCurrentTheme();
     }
 
     changeTheme( theme: string ){
-        //Create URL with new theme and apply to theme id
-        const url = `./assets/css/colors/${ theme }.css`
-        this.linkTheme?.setAttribute('href', url);
-        //save in local storage
-        localStorage.setItem('theme', url);
-        this.checkCurrentTheme();
+        //Update new theme and check
+        this.settingsService.changeTheme( theme );
     }
 
-    checkCurrentTheme(){
-        
-        this.links.forEach( element =>{
-            //to deselect all the items
-            element.classList.remove('working');
-            //get theme name
-            const btnTheme = element.getAttribute('data-theme');
-            //build url
-            const themeUrl = `./assets/css/colors/${ btnTheme }.css`;
-            //get current theme
-            const currentTheme = this.linkTheme?.getAttribute('href');
-            // If both are same then add working class
-            if(themeUrl === currentTheme){
-                element.classList.add('working');
-            }
-        })
-    }
+    
 }
